% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaveInputs.R
\name{cleave}
\alias{cleave}
\title{Server side functionality}
\usage{
cleave(session, selector, inputOptions)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}}

\item{selector}{CSS Selector specifying the input/s}

\item{inputOptions}{List of options passed to Cleave as JSON; see notes}
}
\description{
Cleave inputs from the server
}
\note{
See \url{http://nosir.github.io/cleave.js/} for options

Javascript functions passed to inputType should be wrapped in \code{I()}
}
